TS: Extract Taxonomy/List
